+++ app/components/screens/reservation/timeTable/ReservationTimeTable.tsx
@@ -1,6 +1,6 @@
 import React, {useState} from "react";
 import {View, Text} from "react-native";
-import ReservationTimeTableHeader from "./ReservationTimeTableHeader";
+import Header from "./Header";
 
 interface pickerValueInteface {
 	label: string, // "1"->이번주, "2"->다음주
@@ -17,7 +17,7 @@ function ReservationTimeTable({navigation}) {
 
 	return (
 		<View>
-			<ReservationTimeTableHeader
+			<Header
 				btnListener={reserveBtnListener}
 				pickerValue={pickerValue}
 				setPickerValue={setPickerValue}
